Refactoring gulpfile;  - Putting the gulp tasks into separate files (per task) for better maintainability;
